chore: pr comments, new hoc withSidebarAnnotations
